Fix : restore BackEnd source

+++ server/itclips/src/main/java/com/ssafy/itclips/bookmark/service/BookmarkServiceImpl.java
@@ -1,7 +1,7 @@
 package com.ssafy.itclips.bookmark.service;
 
 import com.ssafy.itclips.bookmark.dto.BookmarkRequestDTO;
-import com.ssafy.itclips.bookmark.dto.BookmarkUpdateDTO;
+import com.ssafy.itclips.global.gpt.GPTResponseDTO;
 import com.ssafy.itclips.bookmark.entity.Bookmark;
 import com.ssafy.itclips.bookmark.entity.BookmarkLike;
 import com.ssafy.itclips.bookmark.repository.BookmarkLikeRepository;
@@ -14,6 +14,8 @@ import com.ssafy.itclips.category.entity.Category;
 import com.ssafy.itclips.category.repository.CategoryRepository;
 import com.ssafy.itclips.error.CustomException;
 import com.ssafy.itclips.error.ErrorCode;
+import com.ssafy.itclips.global.gpt.ChatGPTRequest;
+import com.ssafy.itclips.global.gpt.ChatGPTResponse;
 import com.ssafy.itclips.tag.entity.BookmarkTag;
 import com.ssafy.itclips.tag.entity.Tag;
 import com.ssafy.itclips.tag.repository.BookmarkTagRepository;
@@ -21,14 +23,20 @@ import com.ssafy.itclips.tag.service.TagService;
 import com.ssafy.itclips.user.entity.User;
 import com.ssafy.itclips.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.client.RestTemplate;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class BookmarkServiceImpl implements BookmarkService {
 
     private final BookmarkRepository bookmarkRepository;
@@ -39,34 +47,75 @@ public class BookmarkServiceImpl implements BookmarkService {
     private final CategoryRepository categoryRepository;
     private final BookmarkCategoryRepository bookmarkCategoryRepository;
     private final BookmarkTagRepository bookmarkTagRepository;
+    private final RestTemplate template;
+
+    @Value("${openai.model}")
+    private String model;
+
+    @Value("${openai.api.url}")
+    private String apiURL;
 
     @Override
     @Transactional
-    public void createBookmark(Long listId, Long categoryId, BookmarkRequestDTO bookmarkRequestDTO) throws RuntimeException {
-        BookmarkList existingBookmarkList = bookmarkListRepository.findById(listId)
-                .orElseThrow(() -> new CustomException(ErrorCode.BOOKMARK_LIST_NOT_FOUND));
-        Category existingCategory = categoryRepository.findById(categoryId)
-                .orElseThrow(() -> new CustomException(ErrorCode.CATEGORY_NOT_FOUND));
-        Integer count = bookmarkCategoryRepository.countByCategoryId(categoryId);
-        //dto entity 넣기
+    public void createBookmark(Long userId, Long listId, Long categoryId, BookmarkRequestDTO bookmarkRequestDTO) {
+        BookmarkList existingBookmarkList = getExistingBookmarkList(listId);
+
+        if(!checkAuthority(userId,listId)) {
+            throw new CustomException(ErrorCode.UNAUTHORIZED_USER);
+        }
+
+        Integer count = getBookmarkCount(categoryId);
+
         Bookmark bookmark = buildBookmark(bookmarkRequestDTO, count);
         bookmark.addBookmarkList(existingBookmarkList);
         bookmarkRepository.save(bookmark);
 
-        BookmarkCategory bookmarkCategory = new BookmarkCategory();
-        bookmarkCategory.addBookmarkCategory(bookmark,existingCategory);
-        bookmarkCategoryRepository.save(bookmarkCategory);
+        if (categoryId != null) {
+            Category existingCategory = getExistingCategory(categoryId);
+            saveBookmarkCategory(bookmark, existingCategory);
+        }
 
         saveBookmarkTags(bookmarkRequestDTO, bookmark);
     }
 
+    private Boolean checkAuthority(Long userId, Long listId) {
+        Set<Long> groupUser = bookmarkListRepository.findGroupUserByListId(listId);
+        return groupUser.contains(userId);
+    }
+
+    private BookmarkList getExistingBookmarkList(Long listId) {
+        return bookmarkListRepository.findById(listId)
+                .orElseThrow(() -> new CustomException(ErrorCode.BOOKMARK_LIST_NOT_FOUND));
+    }
+
+    private Integer getBookmarkCount(Long categoryId) {
+        if (categoryId == null) {
+            return 1;
+        }
+        return bookmarkCategoryRepository.countByCategoryId(categoryId);
+    }
+
+    private Category getExistingCategory(Long categoryId) {
+        return categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new CustomException(ErrorCode.CATEGORY_NOT_FOUND));
+    }
+
+    private void saveBookmarkCategory(Bookmark bookmark, Category existingCategory) {
+        BookmarkCategory bookmarkCategory = new BookmarkCategory();
+        bookmarkCategory.addBookmarkCategory(bookmark, existingCategory);
+        bookmarkCategoryRepository.save(bookmarkCategory);
+    }
+
     @Override
     @Transactional
-    public void updateBookmark(Long bookmarkId, BookmarkRequestDTO bookmarkRequestDTO) throws RuntimeException {
+    public void updateBookmark(Long userId, Long bookmarkId, BookmarkRequestDTO bookmarkRequestDTO) throws RuntimeException {
         Bookmark bookmark = bookmarkRepository.findById(bookmarkId)
                 .orElseThrow(() -> new CustomException(ErrorCode.BOOKMARK_NOT_FOUND));
-        bookmark.updateBookmark(bookmarkRequestDTO);
 
+        if(!checkAuthority(userId,bookmark.getBookmarklist().getId())) {
+            throw new CustomException(ErrorCode.UNAUTHORIZED_USER);
+        }
+        bookmark.updateBookmark(bookmarkRequestDTO);
         bookmarkRepository.save(bookmark);
         bookmarkTagRepository.deleteAllByBookmark(bookmark);
 
@@ -82,9 +131,12 @@ public class BookmarkServiceImpl implements BookmarkService {
 
     @Override
     @Transactional
-    public void deleteBookmark(Long bookmarkId) throws RuntimeException {
+    public void deleteBookmark(Long userId, Long bookmarkId) throws RuntimeException {
         Bookmark bookmark = bookmarkRepository.findById(bookmarkId)
                 .orElseThrow(() -> new CustomException(ErrorCode.BOOKMARK_NOT_FOUND));
+        if(!checkAuthority(userId,bookmark.getBookmarklist().getId())) {
+            throw new CustomException(ErrorCode.UNAUTHORIZED_USER);
+        }
         bookmarkRepository.deleteById(bookmarkId);
     }
 
@@ -116,6 +168,20 @@ public class BookmarkServiceImpl implements BookmarkService {
        bookmarkLikeRepository.delete(bookmarkLike);
     }
 
+    @Override
+    @Cacheable(value = "bookmarks", key="#bookmarkId")
+    public GPTResponseDTO getUrlSummary(Long bookmarkId) throws RuntimeException {
+        Bookmark bookmark = bookmarkRepository.findById(bookmarkId)
+                .orElseThrow(() -> new CustomException(ErrorCode.BOOKMARK_NOT_FOUND));
+        String prompt = bookmark.getUrl() + "을 유효한 url이라면 다른 말과 번호 없이 개조식으로 첫 글자에 '-'를 추가해서 3줄로 요약 해 줘. " +
+                "유효하지 않은 url이거나 동영상 사이트라면 '요약할 수 없는 url입니다.' 이 내용만 출력해줘. ";
+        ChatGPTRequest request = new ChatGPTRequest(model,prompt);
+        ChatGPTResponse response = template.postForObject(apiURL, request, ChatGPTResponse.class);
+        if(response == null) {
+            throw new CustomException(ErrorCode.BOOKMARK_SUMMARY_FAILED);
+        }
+        return GPTResponseDTO.of(response);
+    }
 
     private void createTags(BookmarkRequestDTO bookmarkRequestDTO, Bookmark bookmark, List<BookmarkTag> bookmarkTags) {
         List<Tag> tags = tagService.saveTags(bookmarkRequestDTO.getTags());
